requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
tqdm>=4.62.0
opencv-python>=4.5.0
huggingface_hub>=0.16.0
accelerate>=0.20.0 